+++ tests/plugins/image_classification/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2026 Pointmatic
+# SPDX-License-Identifier: Apache-2.0